mcp-server/requirements.txt
mcp[cli]==1.6.0 
openai==1.75.0
python-dotenv
ipykernel
httpx
uv
